backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.0.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
dj-database-url==2.3.0
Django==4.2.17
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-df-api-drf==1.0.4
django-df-chat==0.1.7
django-environ==0.11.2
django-filter==24.3
django-model-utils==5.0.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
gunicorn==23.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
msgpack==1.1.0
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.3.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.33.0
whitenoise==6.8.2
zope.interface==7.2
